drm/i915/tgl: increase number of engines to include CCS
With the compute class/engine already defined, increase the number of
engines to start exposing CCS to other functions.

Note: RCS/CCS engine reset will require to preempt-to-idle the non-hung
engine before the reset can start, because some shared modules will be
reset.

v2: Hold on exposing user-rings (Trvtko)

BSpec: 19212.
Cc: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>

+++ drivers/gpu/drm/i915/intel_device_info.h
@@ -144,7 +144,7 @@ struct sseu_dev_info {
 	u8 eu_mask[GEN_MAX_SLICES * GEN_MAX_SUBSLICES];
 };
 
-typedef u8 intel_ring_mask_t;
+typedef u16 intel_ring_mask_t;
 
 struct intel_device_info {
 	u16 device_id;
